services/collection/requirements.txt
# Collection Service Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Models
-e ../../libs/python/ladon-models

# Google Cloud
google-cloud-pubsub==2.18.4
google-cloud-bigquery==3.13.0
google-cloud-storage==2.13.0

# Database Clients
trino==0.327.0

# HTTP Client
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
